lots of code review comments were addressed
